+++ src/mirakuru/base.py
@@ -37,7 +37,7 @@ from mirakuru.exceptions import (
 )
 from mirakuru.compat import SIGKILL
 
-log = logging.getLogger(__name__)
+log = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
 
 ENV_UUID = 'mirakuru_uuid'
@@ -45,10 +45,15 @@ ENV_UUID = 'mirakuru_uuid'
 Name of the environment variable used by mirakuru to mark its subprocesses.
 """
 
+IGNORED_ERROR_CODES = [errno.ESRCH]
+if platform.system() == 'Darwin':
+    IGNORED_ERROR_CODES = [errno.ESRCH, errno.EPERM]
+
 
 @atexit.register
 def cleanup_subprocesses():
     """On python exit: find possibly running subprocesses and kill them."""
+    # pylint: disable=redefined-outer-name, reimported
     # atexit functions tends to loose global imports sometimes so reimport
     # everything what is needed again here:
     import os
@@ -69,8 +74,8 @@ class SimpleExecutor(object):
     """Simple subprocess executor with start/stop/kill functionality."""
 
     def __init__(
-        self, command, shell=False, timeout=None, sleep=0.1,
-        sig_stop=signal.SIGTERM, sig_kill=SIGKILL
+            self, command, shell=False, timeout=None, sleep=0.1,
+            sig_stop=signal.SIGTERM, sig_kill=SIGKILL
     ):
         """
         Initialize executor.
@@ -240,7 +245,7 @@ class SimpleExecutor(object):
             try:
                 os.kill(pid, sig)
             except OSError as err:
-                if err.errno == errno.ESRCH:
+                if err.errno in IGNORED_ERROR_CODES:
                     # the process has died before we tried to kill it.
                     pass
                 else:
@@ -273,7 +278,7 @@ class SimpleExecutor(object):
         try:
             os.killpg(self.process.pid, sig)
         except OSError as err:
-            if err.errno == errno.ESRCH:
+            if err.errno in IGNORED_ERROR_CODES:
                 pass
             else:
                 raise