+++ tests/executors/test_tcp_executor.py
@@ -9,7 +9,7 @@ from mirakuru.compat import http_server_cmd
 
 PORT = 7986
 
-http_server_cmd = '{0} {1}'.format(http_server_cmd, PORT)
+HTTP_SERVER = '{0} {1}'.format(http_server_cmd, PORT)
 
 
 @pytest.mark.parametrize('timeout', (None, 5))
@@ -40,8 +40,8 @@ def test_it_raises_error_on_timeout():
 
 def test_fail_if_other_executor_running():
     """Test raising AlreadyRunning exception."""
-    executor = TCPExecutor(http_server_cmd, host='localhost', port=PORT)
-    executor2 = TCPExecutor(http_server_cmd, host='localhost', port=PORT)
+    executor = TCPExecutor(HTTP_SERVER, host='localhost', port=PORT)
+    executor2 = TCPExecutor(HTTP_SERVER, host='localhost', port=PORT)
 
     with executor: