+++ src/mirakuru/__init__.py
@@ -33,7 +33,7 @@ from mirakuru.exceptions import (
     ProcessExitedWithError,
 )
 
-__version__ = '0.9.0'
+__version__ = '1.0.0'
 
 __all__ = (
     'Executor',